pyproject.toml
[project]
name = "sudo-sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai",
    "transformers",
    "torch",
    "pyyaml",
]
